90092: Port the class to a model
