+++ modules/pensions/lib/pensions/engine.rb
@@ -16,11 +16,11 @@ module Pensions
       end
     end
 
-    initializer 'pensions.military_information' do |app|
-      app.config.to_prepare do
-        require_relative '../pensions/military_information'
-      end
-    end
+    # initializer 'pensions.military_information', after: :initialize do |app|
+    #   app.config.after_initialize do
+    #     require_relative '../pensions/military_information'
+    #   end
+    # end
 
     initializer 'pensions.pdf_fill.register_form' do |app|
       app.config.to_prepare do
